change add information folder name to resume and update link at the nav
