--- src/components/information-input/InformationInput.jsx
@@ -1,15 +1,0 @@
-import StepperComponents from "./stepper/StepperComponents";
-
-const InformationInput = () => {
-  return (
-    <>
-      <div className="">
-        <div className="max-w-7xl mx-auto pt-30">
-          <StepperComponents />
-        </div>
-      </div>
-    </>
-  );
-};
-
-export default InformationInput;
