+++ src/components/resume/Resume.jsx
@@ -1,0 +1,15 @@
+import StepperComponents from "./stepper/StepperComponents";
+
+const InformationInput = () => {
+  return (
+    <>
+      <div className="">
+        <div className="max-w-7xl mx-auto pt-30">
+          <StepperComponents />
+        </div>
+      </div>
+    </>
+  );
+};
+
+export default InformationInput;
